requirements.txt
acres==0.5.0
aiofiles==24.1.0
anyio==4.10.0
asn1crypto==1.5.1
attr==0.3.2
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
certvalidator==0.11.1
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.1
customtkinter==5.2.2
darkdetect==0.8.0
Deprecated==1.2.18
etelemetry==0.3.1
filelock==3.19.1
fitz==0.0.1.dev2
frontend==0.0.3
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
httplib2==0.31.0
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
looseversion==1.3.0
lxml==6.0.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.3
oscrypto==1.3.0
packaging==25.0
pandas==2.3.2
paramiko==4.0.0
pathlib==1.0.1
pikepdf==9.11.0
pillow==11.3.0
plyer==2.1.0
prov==2.1.1
puremagic==1.30
pycparser==2.23
pydot==4.0.1
PyKCS11==1.5.18
PyMuPDF==1.26.6
PyNaCl==1.6.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyobjus==1.2.3
pyparsing==3.2.4
PyPDF2==3.0.1
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
pystray==0.19.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
qrcode==8.2
rdflib==7.2.1
reportlab==4.4.3
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
shiboken6==6.9.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tools==1.0.5
traits==7.0.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket==0.2.1
websockets==15.0.1
wrapt==1.17.3
zope.event==6.0
zope.interface==8.0
